Result

Description

Result and state handling

Since: 0.1.0

Documentation

type Result = [Benchmark]

The result of the complete run

amount :: Step -> Int

Retrieve the amount of benchmark results for the provided Step

initParserStep :: Step

Initial parser step for convenience

parseStepIO :: Step -> String -> IO Step

Go one step forward and log output

save :: Step -> Environment -> IO ()

Sen the provided data to the given url including the environment